refactor widget
