+++ src/widgets/glances/components/container.jsx
@@ -1,0 +1,8 @@
+export default function Container({ children }) {
+  return (
+    <div>
+      {children}
+      <div className="h-[68px] overflow-clip" />
+    </div>
+  );
+}
