+++ src/widgets/glances/components/custom_tooltip.jsx
